hal/gaisler: extract ambapp to common hal interface
JIRA: RTOS-884
