+++ hal/gaisler/Makefile
@@ -1,0 +1,7 @@
+#
+# Makefile for HAL Gaisler functions
+#
+# Copyright 2024 Phoenix Systems
+#
+
+OBJS += $(addprefix $(PREFIX_O)hal/gaisler/, ambapp.o)
