+++ hal/gaisler/ambapp.c
