+++ hal/gaisler/ambapp.h
