+++ hal/riscv64/noelv/Makefile
@@ -4,4 +4,6 @@
 # Copyright 2024 Phoenix Systems
 #
 
-OBJS += $(addprefix $(PREFIX_O)hal/$(TARGET_SUFF)/$(TARGET_SUBFAMILY)/, ambapp.o console.o noelv.o)
+include hal/gaisler/Makefile
+
+OBJS += $(addprefix $(PREFIX_O)hal/$(TARGET_SUFF)/$(TARGET_SUBFAMILY)/, console.o noelv.o)
